Move SearchContainer to correct folder
This is defined by the search and not the main page, it is only also
used in main page. So this really belongs to /search more than /main.
