+++ ui/src/components/routes-and-lines/search/SearchContainer.tsx
@@ -1,15 +1,15 @@
 import { DateTime } from 'luxon';
 import { useTranslation } from 'react-i18next';
-import { SearchQueryParameterNames, useSearch } from '../../../../hooks';
-import { useToggle } from '../../../../hooks/useToggle';
-import { Column, Container, Row, Visible } from '../../../../layoutComponents';
-import { ChevronToggle, SimpleButton } from '../../../../uiComponents';
-import { AllOptionEnum } from '../../../../utils';
-import { FormRow } from '../../../forms/common';
-import { LineTypeDropdown } from '../../../forms/line/LineTypeDropdown';
-import { VehicleModeDropdown } from '../../../forms/line/VehicleModeDropdown';
-import { PriorityCondition } from './PriorityCondition';
-import { SearchInput } from './SearchInput';
+import { SearchQueryParameterNames, useSearch } from '../../../hooks';
+import { useToggle } from '../../../hooks/useToggle';
+import { Column, Container, Row, Visible } from '../../../layoutComponents';
+import { ChevronToggle, SimpleButton } from '../../../uiComponents';
+import { AllOptionEnum } from '../../../utils';
+import { FormRow } from '../../forms/common';
+import { LineTypeDropdown } from '../../forms/line/LineTypeDropdown';
+import { VehicleModeDropdown } from '../../forms/line/VehicleModeDropdown';
+import { PriorityCondition } from './conditions/PriorityCondition';
+import { SearchInput } from './conditions/SearchInput';
 
 export const SearchContainer = (): JSX.Element => {
   const { searchConditions, setSearchCondition, handleSearch } = useSearch();
